refactored program scraper, moved global variable sinto util files
